refactor: PageLayout 리팩토링
기존 PageLayout을 제거하고, shared/layout.tsx에 새로운 공통 Layout을
추가하였습니다.
