+++ apps/web/app/library/layout.tsx
@@ -1,5 +1,9 @@
-import PageLayout from "@/_components/layout/PageLayout";
+import Layout from "@/_components/shared/layout";
 
 export default function LibraryLayout({ children }: { children: React.ReactNode }) {
-  return <PageLayout>{children}</PageLayout>;
+  return (
+    <Layout top={true} bottom={true} pathname="/library" topbarText="">
+      {children}
+    </Layout>
+  );
 }
