+++ apps/web/app/mypage/layout.tsx
@@ -1,5 +1,9 @@
-import PageLayout from "@/_components/layout/PageLayout";
+import Layout from "@/_components/shared/layout";
 
 export default function MyPageLayout({ children }: { children: React.ReactNode }) {
-  return <PageLayout>{children}</PageLayout>;
+  return (
+    <Layout top={true} bottom={true} pathname="/mypage" topbarText="마이페이지">
+      {children}
+    </Layout>
+  );
 }
